Move C++ `ValType` to its own header
And fill out some small tests
