+++ crates/c-api/include/wasmtime/types.hh
@@ -2,5 +2,6 @@
 #define WASMTIME_TYPES_HH
 
 #include <wasmtime/types/memory.hh>
+#include <wasmtime/types/val.hh>
 
 #endif // WASMTIME_TYPES_HH
